@@ -46,9 +46,9 @@
     "dev:server": "node ./test/adhoc/server.js",
     "lint": "standard && stylelint '**/*.scss'",
     "lint:fix": "standard --fix && stylelint --fix '**/*.scss'",
-    "test": "node ./bin/buildStylesForJest.js && jest --runInBand",
+    "test": "node ./bin/buildStylesForJest.js && NODE_OPTIONS=--experimental-vm-modules jest --runInBand",
     "test:adhoc": "node ./test/adhoc/server.js",
-    "cover": "node ./bin/buildStylesForJest.js && jest --runInBand --coverage",
+    "cover": "node ./bin/buildStylesForJest.js && NODE_OPTIONS=--experimental-vm-modules jest --runInBand --coverage",
     "docs": "node bin/processCustomEmoji.js",
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s",
     "version": "run-s changelog docs && git add CHANGELOG.md docs"
@@ -122,8 +122,8 @@
     "stylelint": "^15.10.3",
     "stylelint-config-recommended-scss": "^13.0.0",
     "stylelint-scss": "^5.2.1",
-    "svelte": "^3.59.1",
-    "svelte-jester": "^2.3.2",
+    "svelte": "^4.2.1",
+    "svelte-jester": "^3.0.0",
     "svelte-preprocess": "^5.0.4",
     "svgo": "^3.0.2",
     "tachometer": "^0.7.0",
